# [CH] optimize hud_query by eliminating joins