Utilizing the new dict + aliased column approach to eliminate expensive join.

perf

```./clickhouse_query_perf.py --query hud_query --perf --times 5 --base HEAD
Using unstaged changes for --head
Gathering perf stats for: hud_query
Num tests: 5
Num times: 5
Waiting for metrics to populate, please be patient
+------+----------+-----------+-------------+---------------+----------+-----------+------------+--------------+
| Test | Avg Time | Base Time | Time Change | % Time Change | Avg Mem  |  Base Mem | Mem Change | % Mem Change |
+------+----------+-----------+-------------+---------------+----------+-----------+------------+--------------+
|  0   |   134    |    415    |     -281    |      -68      | 38258808 | 148456303 | -110197495 |     -74      |
|  1   |   155    |    347    |     -192    |      -55      | 26351850 | 148452495 | -122100645 |     -82      |
|  2   |   130    |    298    |     -168    |      -56      | 52304420 | 148453247 | -96148827  |     -65      |
|  3   |   133    |    309    |     -176    |      -57      | 48468858 | 148452607 | -99983749  |     -67      |
|  4   |   320    |    994    |     -674    |      -68      | 40857464 | 148481827 | -107624363 |     -72      |
+------+----------+-----------+-------------+---------------+----------+-----------+------------+--------------+

```


### Test

```
./clickhouse_query_perf.py --query hud_query --results --times 1  --base HEAD --strict-results
Using unstaged changes for --head
Comparing results for query: hud_query
Num tests: 5
Head: unstaged
 Base: HEAD
Results for test 0 match
Results for test 1 match
Results for test 2 match
Results for test 3 match
Results for test 4 match
```